+++ lib/stoplight/notifier/standard_error.rb
@@ -3,14 +3,14 @@
 module Stoplight
   module Notifier
     class StandardError < Base
-      DEFAULT_FORMAT = '%s'
+      DEFAULT_FORMAT = 'Switching %s from %s to %s'
 
       def initialize(format = nil)
         @format = format || DEFAULT_FORMAT
       end
 
-      def notify(message)
-        warn(@format % message)
+      def notify(light, from_color, to_color)
+        warn(format(@format, light.name, from_color, to_color))
       end
     end
   end